// src/types/cartMenu.ts
import { MenuType } from "./website";

export type CartMenuType = {
  menu: MenuType;
  quantity: number;
};
// export type CartMenuListType = {
//   [menuId: string]: CartMenuType;
// };
// export type CartMenuListType = CartMenuType[];
// export type CartMenuListType = Record<string, CartMenuType>;